+++ src/main/java/com/kelseyde/calvin/tables/history/ContinuationHistoryTable.java
@@ -8,24 +8,24 @@ import com.kelseyde.calvin.engine.EngineConfig;
 
 public class ContinuationHistoryTable extends AbstractHistoryTable {
 
-    int[][][][][] table = new int[2][Piece.COUNT][Square.COUNT][Piece.COUNT][Square.COUNT];
+    short[][][][][] table = new short[2][Piece.COUNT][Square.COUNT][Piece.COUNT][Square.COUNT];
 
     public ContinuationHistoryTable(EngineConfig config) {
-        super(config.contHistBonusMax.value,
-                config.contHistBonusScale.value,
-                config.contHistMalusMax.value,
-                config.contHistMalusScale.value,
-                config.contHistMaxScore.value);
+        super((short) config.contHistBonusMax.value,
+                (short) config.contHistBonusScale.value,
+                (short) config.contHistMalusMax.value,
+                (short) config.contHistMalusScale.value,
+                (short) config.contHistMaxScore.value);
     }
 
     public void update(Move prevMove, Piece prevPiece, Move currMove, Piece currPiece, int depth, boolean white, boolean good) {
-        int current = get(prevMove, prevPiece, currMove, currPiece, white);
-        int bonus = good ? bonus(depth) : malus(depth);
-        int update = gravity(current, bonus);
+        short current = get(prevMove, prevPiece, currMove, currPiece, white);
+        short bonus = good ? bonus(depth) : malus(depth);
+        short update = gravity(current, bonus);
         set(prevMove, prevPiece, currMove, currPiece, update, white);
     }
 
-    public int get(Move prevMove, Piece prevPiece, Move currMove, Piece currPiece, boolean white) {
+    public short get(Move prevMove, Piece prevPiece, Move currMove, Piece currPiece, boolean white) {
         if (prevMove == null || prevPiece == null || currMove == null || currPiece == null) {
             return 0;
         }
@@ -33,7 +33,7 @@ public class ContinuationHistoryTable extends AbstractHistoryTable {
         return table[colourIndex][prevPiece.index()][prevMove.to()][currPiece.index()][currMove.to()];
     }
 
-    public void set(Move prevMove, Piece prevPiece, Move currMove, Piece currPiece, int update, boolean white) {
+    public void set(Move prevMove, Piece prevPiece, Move currMove, Piece currPiece, short update, boolean white) {
         if (prevMove == null || prevPiece == null || currMove == null || currPiece == null) {
             return;
         }
@@ -42,7 +42,7 @@ public class ContinuationHistoryTable extends AbstractHistoryTable {
     }
 
     public void clear() {
-        table = new int[2][Piece.COUNT][Square.COUNT][Piece.COUNT][Square.COUNT];
+        table = new short[2][Piece.COUNT][Square.COUNT][Piece.COUNT][Square.COUNT];
     }
 
 }
